pci driver coming along

+++ include/kernel/io_port.h
@@ -22,9 +22,9 @@ static inline void outl(u32 port, u32 data)
 	asm volatile("outl %0,%w1" : : "a" (data), "d" (port));
 }
 
-static inline u8 inl(u32 port)
+static inline u32 inl(u32 port)
 {
-	u8 data;
+	u32 data;
 	asm volatile("inl %w1,%0" : "=a" (data) : "d" (port));
 	return data;
 }
